refactore TaskBoard component and move to smart container
